Finish refactoring in network pkg
